split githubPermissions into separated components
