+++ packages/app/src/app/components/CreateSandbox/Import/Import.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { useWorkspaceLimits } from 'app/hooks/useWorkspaceLimits';
 import { useWorkspaceSubscription } from 'app/hooks/useWorkspaceSubscription';
 import { useGitHuPermissions } from 'app/hooks/useGitHubPermissions';
-import { RestrictedPublicReposImport } from 'app/pages/Dashboard/Components/shared/githubPermissions';
+import { RestrictedPublicReposImport } from 'app/pages/Dashboard/Components/shared/RestrictedPublicReposImport';
 import { GithubRepoToImport } from './types';
 import { useGithubRepo } from './useGithubRepo';
 import { getOwnerAndRepoFromInput } from './utils';
